Phase 3: Timeline
